+++ lib/src/models/chat/chat_feedback_info.dart
@@ -1,0 +1,12 @@
+class ChatFeedbackInfo {
+  final String authorId;
+  final int rating;
+  final String? comment;
+
+  ChatFeedbackInfo({required this.authorId, required this.rating, this.comment});
+
+  ChatFeedbackInfo.fromJson(Map<String, dynamic> json)
+      : authorId = json['authorId'],
+        rating = json['rating'],
+        comment = json['comment'];
+}