+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2024-01-10
+### Added
+- `ChatMessageInfo.ChatFeedbackInfo` to include Explorer feedback (#103)
+- `Message.messageId` (#103)
+- `PlatformClient.sendChatMessageFeedback` to send user feedback to the AI response (#103)
+
 ## [1.11.0] - 2024-01-03
 ### Added
 - `User.aiDailyMessageLimit` to determine how many AI image description requests an Explorer can make per day (#102)